IDE is being weird
IDE complains about it so it adds it, then complains about it so it wants it removed

+++ src/client/java/com/incrementalclient/common/utils/DurationEstimator.java
@@ -104,7 +104,7 @@ public class DurationEstimator {
         if(denominator == 0) {
             return 0.2f;
         }
-        return (float) (numerator/denominator);
+        return numerator/denominator;
     }
 
     public Duration getEstimate() {
